fix(zone.js): allow draining microtasks in `Promise.then` (through flag)
These changes are essentially the same as those introduced in
https://github.com/angular/angular/pull/45273, but they include backward compatibility
for applications that explicitly rely on the order in which microtasks are drained.

With this change, we introduce a global configuration flag,
`__zone_symbol__enable_native_microtask_draining`, to allow consumers to enable
microtask draining within a browser microtask.

This flag is necessary to prevent any breaking changes resulting from this modification.
The previous attempt to address this issue caused a significant number of failures in g3.
Therefore, we are hiding that fix behind the configuration flag.

Closes #44446